feat: rename components folder
