+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { HashRouter, Route, Routes } from 'react-router-dom';
-import Home from './pages/Home';
-import NotFound from './pages/NotFound';
+import Home from './components/Home';
+import NotFound from './components/NotFound';
 import './App.css';
 
 export function App() {
